+++ nion/swift/test/DataItem_test.py
@@ -786,6 +786,23 @@ class TestDataItemClass(unittest.TestCase):
                 display_item.remove_graphic(crop_region)
                 self.assertTrue(graphics_changed_ref[0])
 
+    def test_changing_calibration_updates_interval_graphics(self):
+        graphics_changed_ref = [False]
+        def graphics_changed(s):
+            graphics_changed_ref[0] = True
+        document_model = DocumentModel.DocumentModel()
+        with contextlib.closing(document_model):
+            data_item = DataItem.DataItem(numpy.zeros((8,)))
+            document_model.append_data_item(data_item)
+            display_item = document_model.get_display_item_for_data_item(data_item)
+            with contextlib.closing(display_item.graphics_changed_event.listen(graphics_changed)):
+                graphic = Graphics.IntervalGraphic()
+                display_item.add_graphic(graphic)
+                self.assertTrue(graphics_changed_ref[0])
+                graphics_changed_ref[0] = False
+                data_item.set_dimensional_calibration(0, Calibration.Calibration(1, 2, "g"))
+                self.assertTrue(graphics_changed_ref[0])
+
     def test_connecting_data_source_updates_dependent_data_items_property_on_source(self):
         document_model = DocumentModel.DocumentModel()
         with contextlib.closing(document_model):